Moved DataObjectAnnotatedInfo & JsonCodecInfo
