+++ src/main/java/io/vertx/codegen/type/DataObjectTypeInfo.java
@@ -14,92 +14,6 @@ import java.util.function.Function;
  */
 public class DataObjectTypeInfo extends ClassTypeInfo {
 
-  public static class JsonCodecInfo {
-    private final String jsonEncoderSimpleName;
-    private final String jsonEncoderPackage;
-    private final String jsonEncoderEnclosingClass;
-
-    private final String jsonDecoderSimpleName;
-    private final String jsonDecoderEnclosingClass;
-    private final String jsonDecoderPackage;
-
-    public JsonCodecInfo(String jsonEncoderSimpleName, String jsonEncoderPackage, String jsonEncoderEnclosingClass, String jsonDecoderSimpleName, String jsonDecoderPackage, String jsonDecoderEnclosingClass) {
-      this.jsonEncoderSimpleName = jsonEncoderSimpleName;
-      this.jsonEncoderPackage = jsonEncoderPackage;
-      this.jsonEncoderEnclosingClass = jsonEncoderEnclosingClass;
-
-      this.jsonDecoderSimpleName = jsonDecoderSimpleName;
-      this.jsonDecoderEnclosingClass = jsonDecoderEnclosingClass;
-      this.jsonDecoderPackage = jsonDecoderPackage;
-    }
-
-    public String getJsonEncoderSimpleName() {
-      return jsonEncoderSimpleName;
-    }
-
-    public String getJsonEncoderPackage() {
-      return jsonEncoderPackage;
-    }
-
-    public String getJsonEncoderEnclosingClass() {
-      return jsonEncoderEnclosingClass;
-    }
-
-    public String getJsonEncoderFQCN() {
-      if (jsonEncoderSimpleName != null)
-        if (jsonEncoderEnclosingClass != null)
-          return jsonEncoderPackage + "." + jsonEncoderEnclosingClass + "." + jsonEncoderSimpleName;
-        else
-          return jsonEncoderPackage + "." + jsonEncoderSimpleName;
-      else return null;
-    }
-
-    public String getJsonDecoderSimpleName() {
-      return jsonDecoderSimpleName;
-    }
-
-    public String getJsonDecoderEnclosingClass() {
-      return jsonDecoderEnclosingClass;
-    }
-
-    public String getJsonDecoderPackage() {
-      return jsonDecoderPackage;
-    }
-
-    public String getJsonDecoderFQCN() {
-      if (jsonDecoderSimpleName != null)
-        if (jsonDecoderEnclosingClass != null)
-          return jsonDecoderPackage + "." + jsonDecoderEnclosingClass + "." + jsonDecoderSimpleName;
-        else
-          return jsonDecoderPackage + "." + jsonDecoderSimpleName;
-      else return null;
-    }
-  }
-
-  public static class DataObjectAnnotatedInfo {
-    private final boolean isDecodeUsingConstructor;
-    private final boolean encodable;
-    private final boolean decodable;
-
-    public DataObjectAnnotatedInfo(boolean isDecodeUsingConstructor, boolean encodable, boolean decodable) {
-      this.isDecodeUsingConstructor = isDecodeUsingConstructor;
-      this.encodable = encodable;
-      this.decodable = decodable;
-    }
-
-    public boolean isDecodableUsingConstructor() {
-      return isDecodeUsingConstructor;
-    }
-
-    public boolean isEncodable() {
-      return encodable;
-    }
-
-    public boolean isDecodable() {
-      return decodable;
-    }
-  }
-
   private final JsonCodecInfo jsonCodecInfo;
   private final DataObjectAnnotatedInfo dataObjectAnnotatedInfo;
 
@@ -144,10 +58,10 @@ public class DataObjectTypeInfo extends ClassTypeInfo {
   }
 
   public boolean isEncodable() {
-    return match(jsonCodecInfo -> jsonCodecInfo.jsonEncoderSimpleName != null, DataObjectAnnotatedInfo::isEncodable);
+    return match(jsonCodecInfo -> jsonCodecInfo.getJsonEncoderSimpleName() != null, DataObjectAnnotatedInfo::isEncodable);
   }
 
   public boolean isDecodable() {
-    return match(jsonCodecInfo -> jsonCodecInfo.jsonDecoderSimpleName != null, DataObjectAnnotatedInfo::isDecodable);
+    return match(jsonCodecInfo -> jsonCodecInfo.getJsonDecoderSimpleName() != null, DataObjectAnnotatedInfo::isDecodable);
   }
 }
